% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishing_area_data.R
\name{fishing_area_data}
\alias{fishing_area_data}
\title{Extract Data for a Specified Fishing Area}
\usage{
fishing_area_data(area_name)
}
\arguments{
\item{area_name}{The name of the FAO major fishing area.}
}
\value{
A data frame containing data for the specified fishing area.
}
\description{
This function extracts and returns data for a specified FAO major fishing area.
}
\examples{
fishing_area_data("Atlantic, Eastern Central")
}
